+++ spec/unit/braintree/config_spec.js
@@ -3,7 +3,6 @@
 
 let Config = require('../../../lib/braintree/config').Config;
 let Environment = require('../../../lib/braintree/environment').Environment;
-let sinon = require('sinon');
 
 describe('Config', function () {
   it('can be configured with merchant credentials', function () {
@@ -40,7 +39,7 @@ describe('Config', function () {
         environment: 'development'
       });
     }
-      , 'Missing merchantId')
+    , 'Missing merchantId')
   );
 
   it('raises an exception if merchantId is empty', () =>
@@ -52,7 +51,7 @@ describe('Config', function () {
         environment: 'development'
       });
     }
-      , 'Missing merchantId')
+    , 'Missing merchantId')
   );
 
   it('raises an exception if publicKey is null', () =>
@@ -63,7 +62,7 @@ describe('Config', function () {
         environment: 'development'
       });
     }
-      , 'Missing publicKey')
+    , 'Missing publicKey')
   );
 
   it('raises an exception if publicKey is empty', () =>
@@ -75,7 +74,7 @@ describe('Config', function () {
         environment: 'development'
       });
     }
-      , 'Missing publicKey')
+    , 'Missing publicKey')
   );
 
   it('raises an exception if privateKey is null', () =>
@@ -86,7 +85,7 @@ describe('Config', function () {
         environment: 'development'
       });
     }
-      , 'Missing privateKey')
+    , 'Missing privateKey')
   );
 
   it('raises an exception if privateKey is empty', () =>
@@ -98,7 +97,7 @@ describe('Config', function () {
         environment: 'development'
       });
     }
-      , 'Missing privateKey')
+    , 'Missing privateKey')
   );
 
   it('raises an exception if environment is null', () =>
@@ -109,7 +108,7 @@ describe('Config', function () {
         privateKey: 'privateKey'
       });
     }
-      , 'Missing environment')
+    , 'Missing environment')
   );
 
   it('raises an exception if environment is empty', () =>
@@ -121,7 +120,7 @@ describe('Config', function () {
         environment: ''
       });
     }
-      , 'Missing environment')
+    , 'Missing environment')
   );
 
   describe('baseMerchantUrl', () =>
@@ -159,15 +158,18 @@ describe('Config', function () {
       assert.equal(config.environment, Environment.Development);
     });
 
-    it('logs an error if config environment is provided and not the same as accessToken environment', () => {
-      sinon.stub(console, 'error');
+    it('logs an error if config environment is provided and not the same as accessToken environment', done => {
+      try {
+        new Config({
+          accessToken: 'access_token$development$integration_merchant_id$f388b1cc',
+          environment: 'production'
+        });
+      } catch (e) {
+        assert.equal(e.message, 'AccessToken environment does not match environment passed in config');
+        done();
+      }
 
-      new Config({
-        accessToken: 'access_token$development$integration_merchant_id$f388b1cc',
-        environment: 'production'
-      });
-
-      assert.equal(console.error.firstCall.args[0], 'Warning: AccessToken environment does not match environment passed in config'); // eslint-disable-line no-console
+      done(assert.fail('Test should have thrown'));
     });
   });
 });